requirements.txt
django-registration-redux==1.1
django-crispy-forms==1.4.0
django-allauth==0.25.2
